Fixes for 8 GPU full system test configuration

+++ prodtests/full-system-test/start_tmux.sh
@@ -12,21 +12,30 @@ source $MYDIR/setenv.sh
 export NGPUS=4
 export GPUTYPE=HIP
 export SHMSIZE=$(( 128 << 30 ))
+export DDSHMSIZE=$(( 64 << 10 ))
+export GPUMEMSIZE=$(( 24 << 30 ))
 export NUMAGPUIDS=1
 export EXTINPUT=1
 export EPNPIPELINES=1
 export SYNCMODE=1
 export SHMTHROW=0
+export SEVERITY=error
 
 if [ $1 == "dd" ]; then
   export CMD=datadistribution.sh
 else
   export CMD=raw-reader.sh
+  export NTIMEFRAMES=1000000
+fi
+
+if [ ! -f matbud.root -a -f ctf_dictionary.root ]; then
+  echo matbud.root or ctf_dictionary.root missing
+  exit 1
 fi
 
 rm -f /dev/shm/*fmq*
 
 tmux \
-    new-session "NUMAID=0 numactl --membind 0 --cpunodebind 0 $MYDIR/dpl-workflow.sh" \; \
-    split-window "NUMAID=1 numactl --membind 1 --cpunodebind 1 $MYDIR/dpl-workflow.sh" \; \
-    split-window "sleep 30; numactl --cpunodebind 0 --interleave=all $MYDIR/$CMD"
+    new-session "NUMAID=0 numactl --membind 0 --cpunodebind 0 $MYDIR/dpl-workflow.sh; echo END; sleep 1000" \; \
+    split-window "sleep 30; NUMAID=1 numactl --membind 1 --cpunodebind 1 $MYDIR/dpl-workflow.sh; echo END; sleep 1000" \; \
+    split-window "sleep 60; numactl --interleave=all $MYDIR/$CMD; echo END; sleep 1000"
